Added Items from Lecture 11
A bit confused now. As a zoom watcher there was slightly hard to hear chatter about orbits. Unsure if both the Earth AND the Moon are orbiting the sun. OR if the Earth is orbiting the Sun, and then the Moon orbits the Earth. 

Noticed that strict mode is on. Gives a warning since we've done item = item.new without using var/let, but I'm not sure if it affects the actual implementation or not. Just gonna leave it there.