The completed worksheets (probably)
For the bounding volume worksheet I wasn't sure what units were being used or if it's just a choose your own number for a function type of thing. But doing math for what looked like distance and space, without units, felt like eating a bowl of cereal and milk using my hands instead of a spoon. Depending on how it works I might need to redo the numbers.