Somehow things are doing things, if only I knew
Splicing together lecture 11 and lecture 8 I managed to get Earth and the sun to appear. They're not rotating yet, but in forcing color to be blue they appear to exist somewhat correctly. At least the earth is smaller and not sitting in the sun. As an Earthling myself, I consider that a small victory.