Moons over my hammy
Double revolution! Just a temporary fix to see how the pieces tangle with one another. I was able to get the earth AND the moon to appear and do a little dance. The moon isn't scaling right since it's using its base value and is thus twice as big as the earth. But from what I could see through the tears, the moon was slowly rotating around the earth while the earth was rotating around the sun! Primarily I need to fix the moon's scaling, the moon's distance, and the overall color issue. Progress though!