Canvas up, trying to get pic on it
Unsure where to start I'm starting off at the beginning. With a canvas displayed through HTML which I'll be using as the viewer, I'm trying to display an image within the canvas using drawImage. If that works I plan to employ sliders for rotate, zoom and pan. There also seems to be tools for interactively doing so, but I don't want to get ahead of myself since my picture isn't appearing yet.