Added in sliders
With the exam done and hopefully not a complete misery, I'm back on the project. The sliders were surprisingly easy to add in despite how I imagined it. After consulting the documentation sliders are literally an html object type. Using input and event listeners I can now move values from the html to the js file and reflect the change in the html. Which I thought was going to be a big headache but ended up moving along very smoothly, I thank google everyday.