Added initial zoom
This just connects the zoom function, which currently does a redraw to scale the image. I'm sure there's a cleaner way but just using html and basic javascript this seems to be the way to go. At the very least it works and doesn't freeze. I still need to be able to keep track of the current "center" so that zoom functions logically. After this I think the good next step would be to do the panning X and Y sliders, then finish off with the rotation. If time permits then using a more consumable double-click pan and scroll zoom would be preferable but just with the amount of trouble I'm having, as well as the weight of other classes, I don't think I'll be able to get to them in time.