Added User Image Upload Function
With this I'm as far done as I think I can get with the presentation tomorrow. This should allow the user to import their own images to toy around and view. I'm not sure about my grade on the ungraded things, so I really only know it from the midterm and the first repository assignment. But hopefully I managed to scrape together enough points to pass. If by chance you're the professor and reading this: Thanks for the semester, have a great summer my guy!